🚚 Move Inhabitant components to their own folder for clarity
